# src/models/setting.go
package models

import "github.com/uptrace/bun"

type Setting struct {
	bun.BaseModel `bun:"table:settings,alias:st" swaggerignore:"true"`

	Key   string `bun:"key,pk"    json:"key"`
	Value string `bun:"value"     json:"value"`
}